% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BlancoIhleBacktest.R, R/ChristoffersenBacktestForIndependence.R
\name{BlancoIhleBacktest}
\alias{BlancoIhleBacktest}
\title{Blanco-Ihle forecast evaluation backtest measure}
\usage{
BlancoIhleBacktest(Ra, Rb, cl)

BlancoIhleBacktest(Ra, Rb, cl)
}
\arguments{
\item{Ra}{Vector of a portfolio profit and loss}

\item{Rb}{Vector of corresponding VaR forecasts}

\item{cl}{VaR confidence interval}

\item{Rc}{Vector of corresponding Expected Tailed Loss forecasts}

\item{Ra}{Vector of portfolio profit and loss observations}

\item{Rb}{Vector of corresponding VaR forecasts}

\item{cl}{Confidence interval for}
}
\value{
Something

Probability that given the data set, the null hypothesis
(i.e. independence) is correct.
}
\description{
Derives the Blanco-Ihle forecast evaluation loss measure for a VaR
risk measurement model.

Carries out the Christoffersen backtest of independence for a VaR risk
measurement model, for specified VaR confidence level.
}
\examples{
# To be added
# To be added
}
\author{
Dinesh Acharya

Dinesh Acharya
}
\references{
Dowd, Kevin. Measuring Market Risk, Wiley, 2007.

Blanco, C. and Ihle, G. How Good is Your Var? Using Backtesting to Assess
System Performance. Financial Engineering News, 1999.

Dowd, K. Measuring Market Risk, Wiley, 2007.

Christoffersen, P. Evaluating Interval Forecasts. International Economic
Review, 39(4), 1992, 841-862.
}

