% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BootstrapVaRFigure.R
\name{BootstrapVaRFigure}
\alias{BootstrapVaRFigure}
\title{Plots figure of bootstrapped VaR}
\usage{
BootstrapVaRFigure(Ra, number.resamples, cl)
}
\arguments{
\item{Ra}{Vector corresponding to profit and loss distribution}

\item{number.resamples}{Number of samples to be taken in bootstrap procedure}

\item{cl}{Number corresponding to Value at Risk confidence level}
}
\description{
Plots figure for the bootstrapped VaR, for confidence
level and holding period implied by data frequency.
}
\examples{
# To be modified with appropriate data.
   # Estimates 90\% confidence interval for bootstrapped VaR for 95\%
   # confidence interval
   Ra <- rnorm(1000)
   BootstrapESFigure(Ra, 500, 0.95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

