\name{Dowd-package}
\alias{Dowd-package}
\docType{package}
\title{
R-version of Kevin Dowd's MATLAB Toolbox from book "Measuring Market Risk".
}

\description{
\kbd{Dowd} Kevin Dowd's book "Measuring Market Risk" gives overview of risk measurement procedures with focus on Value at Risk (VaR) and Expected Shortfall (ES).
}
\author{
Dinesh Acharya \cr

Maintainer: Dinesh Acharya \email{dines.acharya@gmail.com}
}

\references{

Dowd, K. \emph{Measuring Market Risk}. Wiley. 2005.

}

\section{Acknowledgments}{
Without Kevin Dowd's book Measuring Market Risk and accompanying MATLAB toolbox, this project would not have been possible.

Peter Carl and Brian G. Peterson deserve special acknowledgement for mentoring me on this project.
}

\keyword{ package }