% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FilterStrategyLogNormalVaR.R
\name{FilterStrategyLogNormalVaR}
\alias{FilterStrategyLogNormalVaR}
\title{Log Normal VaR with filter strategy}
\usage{
FilterStrategyLogNormalVaR(mu, sigma, number.trials, alpha, cl, hp)
}
\arguments{
\item{mu}{Mean arithmetic return}

\item{sigma}{Standard deviation of arithmetic return}

\item{number.trials}{Number of trials used in the simulations}

\item{alpha}{Participation parameter}

\item{cl}{Confidence Level}

\item{hp}{Holding Period}
}
\value{
Lognormal VaR
}
\description{
Generates Monte Carlo lognormal VaR with filter portfolio strategy
}
\examples{
# Estimates standard error of normal quantile estimate
   FilterStrategyLogNormalVaR(0, .2, 100, 1.2, .95, 10)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

