% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GParetoMultipleMEFPlot.R
\name{GParetoMultipleMEFPlot}
\alias{GParetoMultipleMEFPlot}
\title{Plot of Emperical and 2 Generalised Pareto mean excess functions}
\usage{
GParetoMultipleMEFPlot(Ra, mu, beta, zeta1, zeta2)
}
\arguments{
\item{Ra}{Vector of daily Profit/Loss data}

\item{mu}{Location parameter}

\item{beta}{Scale parameter}

\item{zeta1}{Assumed tail index for first mean excess function}

\item{zeta2}{Assumed tail index for second mean excess function}
}
\description{
Plots of emperical mean excess function and two generalized pareto mean excess functions which differ in their tail-index value.
}
\examples{
# Computes ES assuming generalised Pareto for following parameters
   Ra <- 5 * rnorm(100)
   mu <- 1
   beta <- 1.2
   zeta1 <- 1.6
   zeta2 <- 2.2
   GParetoMultipleMEFPlot(Ra, mu, beta, zeta1, zeta2)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

