% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GumbelESPlot2DCl.R
\name{GumbelESPlot2DCl}
\alias{GumbelESPlot2DCl}
\title{Gumbel VaR}
\usage{
GumbelESPlot2DCl(mu, sigma, n, cl, hp)
}
\arguments{
\item{mu}{Location parameter for daily L/P}

\item{sigma}{Assumed scale parameter for daily L/P}

\item{n}{size from which the maxima are drawn}

\item{cl}{VaR confidence level}

\item{hp}{VaR holding period}
}
\description{
Estimates the EV VaR of a portfolio assuming extreme losses are Gumbel distributed, for specified confidence level and holding period.
}
\examples{
# Plots ES against Cl
   GumbelESPlot2DCl(0, 1.2, 100, seq(0.8,0.99,0.02), 280)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

