% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GumbelVaR.R
\name{GumbelVaR}
\alias{GumbelVaR}
\title{Gumbel VaR}
\usage{
GumbelVaR(mu, sigma, n, cl, hp)
}
\arguments{
\item{mu}{Location parameter for daily L/P}

\item{sigma}{Assumed scale parameter for daily L/P}

\item{n}{Size from which the maxima are drawn}

\item{cl}{VaR confidence level}

\item{hp}{VaR holding period}
}
\value{
Estimated VaR
}
\description{
Estimates the EV VaR of a portfolio assuming extreme losses are Gumbel distributed, for specified confidence level and holding period.
}
\examples{
# Gumbel VaR
   GumbelVaR(0, 1.2, 100, c(.9,.88, .85, .8), 280)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

