% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GumbelVaRPlot2DCl.R
\name{GumbelVaRPlot2DCl}
\alias{GumbelVaRPlot2DCl}
\title{Gumbel VaR}
\usage{
GumbelVaRPlot2DCl(mu, sigma, n, cl, hp)
}
\arguments{
\item{mu}{Location parameter for daily L/P}

\item{sigma}{Assumed scale parameter for daily L/P}

\item{n}{size from which the maxima are drawn}

\item{cl}{VaR confidence level}

\item{hp}{VaR holding period}
}
\description{
Estimates the EV VaR of a portfolio assuming extreme losses are Gumbel distributed, for specified confidence level and holding period.
}
\examples{
# Plots VaR against Cl
   GumbelVaRPlot2DCl(0, 1.2, 100, c(.9,.88, .85, .8), 280)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

