% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HillPlot.R
\name{HillPlot}
\alias{HillPlot}
\title{Hill Plot}
\usage{
HillPlot(Ra, maximum.tail.size)
}
\arguments{
\item{Ra}{The data set}

\item{maximum.tail.size}{maximum tail size and should be greater than a
quarter of the sample size.}
}
\description{
Displays a plot of the Hill Estimator against tail sample size.
}
\examples{
# Hill Estimator - Tail Sample Size Plot for random normal dataset
   Ra <- rnorm(1000)
   HillPlot(Ra, 180)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

