% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HSESFigure.R
\name{HSESFigure}
\alias{HSESFigure}
\title{Figure of Historical SImulation VaR and ES and histogram of L/P}
\usage{
HSESFigure(Ra, cl)
}
\arguments{
\item{Ra}{Vector of profit loss data}

\item{cl}{VaR confidence level}
}
\description{
Plots figure showing the historical simulation VaR and ES and histogram
 of L/P for specified confidence level and holding period implied by data
 frequency.
}
\examples{
# Plots figure showing VaR and histogram of P/L data
   Ra <- rnorm(100)
   HSESFigure(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

