% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KernelESBoxKernel.R
\name{KernelESBoxKernel}
\alias{KernelESBoxKernel}
\title{Calculates ES using box kernel approach}
\usage{
KernelESBoxKernel(Ra, cl)
}
\arguments{
\item{Ra}{Profit and Loss data set}

\item{cl}{VaR confidence level}
}
\value{
Scalar VaR
}
\description{
The output consists of a scalar ES for specified confidence level.
}
\examples{
# VaR for specified confidence level using box kernel approach
   Ra <- rnorm(30)
   KernelESBoxKernel(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

