% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KernelESEpanechinikovKernel.R
\name{KernelESEpanechinikovKernel}
\alias{KernelESEpanechinikovKernel}
\title{Calculates ES using Epanechinikov kernel approach}
\usage{
KernelESEpanechinikovKernel(Ra, cl, plot = TRUE)
}
\arguments{
\item{Ra}{Profit and Loss data set}

\item{cl}{ES confidence level}

\item{plot}{Bool, plots cdf if true}
}
\value{
Scalar ES
}
\description{
The output consists of a scalar ES for specified confidence level.
}
\examples{
# ES for specified confidence level using Epanechinikov kernel approach
   Ra <- rnorm(30)
   KernelESEpanechinikovKernel(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

