% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KernelESTriangleKernel.R
\name{KernelESTriangleKernel}
\alias{KernelESTriangleKernel}
\title{Calculates ES using triangle kernel approach}
\usage{
KernelESTriangleKernel(Ra, cl)
}
\arguments{
\item{Ra}{Profit and Loss data set}

\item{cl}{VaR confidence level}
}
\value{
Scalar VaR
}
\description{
The output consists of a scalar ES for specified confidence level.
}
\examples{
# VaR for specified confidence level using triangle kernel approach
   Ra <- rnorm(30)
   KernelESTriangleKernel(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

