% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KernelVaRBoxKernel.R
\name{KernelVaRBoxKernel}
\alias{KernelVaRBoxKernel}
\title{Calculates VaR using box kernel approach}
\usage{
KernelVaRBoxKernel(Ra, cl, plot = TRUE)
}
\arguments{
\item{Ra}{Profit and Loss data set}

\item{cl}{VaR confidence level}

\item{plot}{Bool which indicates whether the graph is plotted or not}
}
\value{
Scalar VaR
}
\description{
The output consists of a scalar VaR for specified confidence level.
}
\examples{
# VaR for specified confidence level using box kernel approach
   Ra <- rnorm(30)
   KernelVaRBoxKernel(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

