% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KernelVaRNormalKernel.R
\name{KernelVaRNormalKernel}
\alias{KernelVaRNormalKernel}
\title{Calculates VaR using normal kernel approach}
\usage{
KernelVaRNormalKernel(Ra, cl, plot = TRUE)
}
\arguments{
\item{Ra}{Profit and Loss data set}

\item{cl}{VaR confidence level}

\item{plot}{Bool, plots cdf if true}
}
\value{
Scalar VaR
}
\description{
The output consists of a scalar VaR for specified confidence level.
}
\examples{
# VaR for specified confidence level using normal kernel approach
   Ra <- rnorm(30)
   KernelVaRNormalKernel(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

