% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KernelVaRTriangleKernel.R
\name{KernelVaRTriangleKernel}
\alias{KernelVaRTriangleKernel}
\title{Calculates VaR using triangle kernel approach}
\usage{
KernelVaRTriangleKernel(Ra, cl, plot = TRUE)
}
\arguments{
\item{Ra}{Profit and Loss data set}

\item{cl}{VaR confidence level}

\item{plot}{Bool, plots cdf if true.}
}
\value{
Scalar VaR
}
\description{
The output consists of a scalar VaR for specified confidence level.
}
\examples{
# VaR for specified confidence level using triangle kernel approach
   Ra <- rnorm(30)
   KernelVaRTriangleKernel(Ra, .95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

