% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogtESPlot3D.R
\name{LogtESPlot3D}
\alias{LogtESPlot3D}
\title{Plots log-t ES against confidence level and holding period}
\usage{
LogtESPlot3D(...)
}
\arguments{
\item{returns}{Vector of daily geometric return data}

\item{mu}{Mean of daily geometric return data}

\item{sigma}{Standard deviation of daily geometric return data}

\item{investment}{Size of investment}

\item{df}{Number of degrees of freedom in the t distribution}

\item{cl}{VaR confidence level and must be a vector}

\item{hp}{VaR holding period and must be a vector}
}
\description{
Plots the ES of a portfolio against confidence level and holding period assuming that geometric
 returns are Student-t distributed, for specified confidence level and
 holding period.
}
\note{
The input arguments contain either return data or else mean and
 standard deviation data. Accordingly, number of input arguments is either 5
 or 6. In case there 5 input arguments, the mean and standard deviation of
 data is computed from return data. See examples for details.
}
\examples{
# Plots ES against confidene level given geometric return data
   data <- runif(5, min = 0, max = .2)
   LogtESPlot3D(returns = data, investment = 5, df = 6, cl = seq(.85,.99,.01), hp = 60:90)

   # Computes ES against confidence level given mean and standard deviation of return data
   LogtESPlot3D(mu = .012, sigma = .03, investment = 5, df = 6, cl = seq(.85,.99,.02), hp = 40:80)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

