% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LopezBacktest.R
\name{LopezBacktest}
\alias{LopezBacktest}
\title{First (binomial) Lopez forecast evaluation backtest score measure}
\usage{
LopezBacktest(Ra, Rb, cl)
}
\arguments{
\item{Ra}{Vector of portfolio of profit loss distribution}

\item{Rb}{Vector of corresponding VaR forecasts}

\item{cl}{VaR confidence level}
}
\value{
Something
}
\description{
Derives the first Lopez (i.e. binomial) forecast evaluation score
for a VaR risk measurement model.
}
\examples{
# Has to be modified with appropriate data:
   # LopezBacktest for given parameters
   a <- rnorm(1*100)
   b <- abs(rnorm(1*100))+2
   LopezBacktest(a, b, 0.95)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.

Lopez, J. A. Methods for Evaluating Value-at-Risk Estimates. Federal Reserve
Bank of New York Economic Policy Review, 1998, p. 121.

Lopez, J. A. Regulatory Evaluations of Value-at-Risk Models. Journal of Risk
1999, 37-64.
}

