% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MEFPlot.R
\name{MEFPlot}
\alias{MEFPlot}
\title{Mean Excess Function Plot}
\usage{
MEFPlot(Ra)
}
\arguments{
\item{Ra}{Vector data}
}
\description{
Plots mean-excess function values of the data set.
}
\examples{
# Plots mean-excess function values
   Ra <- rnorm(1000)
   MEFPlot(Ra)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

