% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NormalQQPlot.R
\name{NormalQQPlot}
\alias{NormalQQPlot}
\title{Normal Quantile Quantile Plot}
\usage{
NormalQQPlot(Ra)
}
\arguments{
\item{Ra}{Vector data set}
}
\description{
Produces an emperical QQ-Plot of the quantiles of the data set 'Ra' versus
the quantiles of a normal distribution. The purpose of the quantile-quantile plot is to determine whether the sample in 'Ra' is drawn from a normal (i.e., Gaussian) distribution.
}
\examples{
# Normal QQ Plot for randomly generated standard normal data
   Ra <- rnorm(100)
   NormalQQPlot(Ra)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

