% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NormalQuantileStandardError.R
\name{NormalQuantileStandardError}
\alias{NormalQuantileStandardError}
\title{Standard error of normal quantile estimate}
\usage{
NormalQuantileStandardError(prob, n, mu, sigma, bin.size)
}
\arguments{
\item{prob}{Tail probability. Can be a vector or scalar}

\item{n}{Sample size}

\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the distribution}

\item{bin.size}{Bin size. It is optional parameter with default value 1}
}
\value{
Vector or scalar
depending on whether the probability is a vector
or scalar
}
\description{
Estimates standard error of normal quantile estimate
}
\examples{
# Estimates standard error of normal quantile estimate
   NormalQuantileStandardError(.8, 100, 0, .5, 3)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

