% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCAPrelim.R
\name{PCAPrelim}
\alias{PCAPrelim}
\title{PCAPrelim}
\usage{
PCAPrelim(Ra)
}
\arguments{
\item{Ra}{Matrix return data set where each row is interpreted as a set of
daily observations, and each column as the returns to each position in a
portfolio}

\item{position.data}{Position-size vector, giving amount invested in each
position}
}
\description{
Estimates VaR plot using principal components analysis
}
\examples{
# Computes PCA Prelim
   Ra <- matrix(rnorm(4*4), 4, 4)
   PCAPrelim(Ra)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

