% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCAVaRPlot.R
\name{PCAVaRPlot}
\alias{PCAVaRPlot}
\title{VaR plot}
\usage{
PCAVaRPlot(Ra, position.data)
}
\arguments{
\item{Ra}{Matrix return data set where each row is interpreted as a set of daily observations, and each column as the returns to each position in a portfolio}

\item{position.data}{Position-size vector, giving amount invested in each position}
}
\description{
Estimates VaR plot using principal components analysis
}
\examples{
# Computes PCA VaR
   Ra <- matrix(rnorm(15*20),15,20)
   position.data <- rnorm(20)
   PCAVaRPlot(Ra, position.data)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

