% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PickandsPlot.R
\name{PickandsPlot}
\alias{PickandsPlot}
\title{Pickand Estimator - Tail Sample Size Plot}
\usage{
PickandsPlot(Ra, maximum.tail.size)
}
\arguments{
\item{Ra}{The data set}

\item{maximum.tail.size}{maximum tail size and should be greater than a
quarter of the sample size.}
}
\description{
Displays a plot of the Pickands Estimator against Tail Sample Size.
}
\examples{
# Pickand - Sample Tail Size Plot for random standard normal data
   Ra <- rnorm(1000)
   PickandsPlot(Ra, 40)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

