% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StopLossLogNormalVaR.R
\name{StopLossLogNormalVaR}
\alias{StopLossLogNormalVaR}
\title{Log Normal VaR with stop loss limit}
\usage{
StopLossLogNormalVaR(mu, sigma, number.trials, loss.limit, cl, hp)
}
\arguments{
\item{mu}{Mean arithmetic return}

\item{sigma}{Standard deviation of arithmetic return}

\item{number.trials}{Number of trials used in the simulations}

\item{loss.limit}{Stop Loss limit}

\item{cl}{Confidence Level}

\item{hp}{Holding Period}
}
\value{
Lognormal VaR
}
\description{
Generates Monte Carlo lognormal VaR with stop-loss limit
}
\examples{
# Estimates standard error of normal quantile estimate
   StopLossLogNormalVaR(0, .2, 100, 1.2, .95, 10)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

