% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TQQPlot.R
\name{TQQPlot}
\alias{TQQPlot}
\title{Student's T Quantile - Quantile Plot}
\usage{
TQQPlot(Ra, df)
}
\arguments{
\item{Ra}{Sample data set}

\item{df}{Number of degrees of freedom of the t distribution}
}
\description{
Creates emperical QQ-plot of the quantiles of the data set x versus of a t
distribution. The QQ-plot can be used to determine whether the sample in x is
drawn from a t distribution with specified number of degrees of freedom.
}
\examples{
# t-QQ Plot for randomly generated standard normal data
   Ra <- rnorm(100)
   TQQPlot(Ra, 20)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

