% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tVaRFigure.R
\name{tVaRFigure}
\alias{tVaRFigure}
\title{Figure of t- VaR and pdf against L/P}
\usage{
tVaRFigure(...)
}
\arguments{
\item{...}{The input arguments contain either return data or else mean and
 standard deviation data. Accordingly, number of input arguments is either 4
 or 5. In case there 4 input arguments, the mean and standard deviation of
 data is computed from return data. See examples for details.

 returns Vector of daily geometric return data

 mu Mean of daily geometric return data

 sigma Standard deviation of daily geometric return data

 df Number of degrees of freedom

 cl VaR confidence level and should be scalar

 hp VaR holding period in days and should be scalar}
}
\description{
Gives figure showing the VaR and probability distribution function against L/P
 of a portfolio assuming P/L are normally distributed, for specified
 confidence level and holding period.
}
\examples{
# Plots normal VaR and pdf against L/P data for given returns data
   data <- runif(5, min = 0, max = .2)
   tVaRFigure(returns = data, df = 7, cl = .95, hp = 90)

   # Plots normal VaR and pdf against L/P data with given parameters
   tVaRFigure(mu = .012, sigma = .03, df=7, cl = .95, hp = 90)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

