% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogtESPlot3D.R
\name{LogtESPlot3D}
\alias{LogtESPlot3D}
\title{Plots log-t ES against confidence level and holding period}
\usage{
LogtESPlot3D(...)
}
\arguments{
\item{...}{The input arguments contain either return data or else mean and
 standard deviation data. Accordingly, number of input arguments is either 5
 or 6. In case there 5 input arguments, the mean and standard deviation of
 data is computed from return data. See examples for details.

 returns Vector of daily geometric return data

 mu Mean of daily geometric return data

 sigma Standard deviation of daily geometric return data

 investment Size of investment

 df Number of degrees of freedom in the t distribution

 cl VaR confidence level and must be a vector

 hp VaR holding period and must be a vector}
}
\description{
Plots the ES of a portfolio against confidence level and holding period assuming that geometric
 returns are Student-t distributed, for specified confidence level and
 holding period.
}
\examples{
# Plots ES against confidene level given geometric return data
   data <- rnorm(5, .09, .03)
   LogtESPlot3D(returns = data, investment = 5, df = 6, cl = seq(.9,.99,.01), hp = 1:100)

   # Computes ES against confidence level given mean and standard deviation of return data
   LogtESPlot3D(mu = .012, sigma = .03, investment = 5, df = 6, cl = seq(.9,.99,.01), hp = 1:100)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

