% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCAPrelim.R
\name{PCAPrelim}
\alias{PCAPrelim}
\title{Estimates VaR plot using principal components analysis}
\usage{
PCAPrelim(Ra)
}
\arguments{
\item{Ra}{Matrix return data set where each row is interpreted as a set of
daily observations, and each column as the returns to each position in a
portfolio position}
}
\description{
Estimates VaR plot using principal components analysis
}
\examples{
# Computes PCA Prelim
   # This code was based on Dowd's code and similar to Dowd's code,
   # it is inconsistent for non-scalar data (Ra).
   library(MASS)
   Ra <- .15
   PCAPrelim(Ra)
}
\author{
Dinesh Acharya
}
\references{
Dowd, K. Measuring Market Risk, Wiley, 2007.
}

